Solve Image Loading Problem
1. Solve Image Loading Proble using setTag()
2. Issue : Dream Delete dialog doesn't work well(coners are angled)
