+++ app/src/main/java/com/teamnexters/zaza/ui/dream/DreamItemAdapter.kt
@@ -14,6 +14,21 @@ import java.text.SimpleDateFormat
 import java.util.*
 import kotlin.collections.ArrayList
 import com.teamnexters.zaza.ui.dream.DreamItem as DreamItem
+import android.view.animation.OvershootInterpolator
+import androidx.core.view.ViewCompat.animate
+import android.R.attr.scaleX
+import android.R.attr.scaleY
+import android.graphics.Bitmap
+import android.graphics.drawable.BitmapDrawable
+import android.graphics.drawable.Drawable
+import androidx.core.view.ViewCompat.setScaleY
+import androidx.core.view.ViewCompat.setScaleX
+import androidx.core.view.ViewCompat.setAlpha
+import android.text.TextUtils
+import com.squareup.picasso.Picasso
+import com.squareup.picasso.Target
+import java.lang.Exception
+
 
 class DreamItemAdapter : RecyclerView.Adapter<DreamItemViewHolder> {
     private var items: ArrayList<DreamItem>
@@ -38,6 +53,26 @@ class DreamItemAdapter : RecyclerView.Adapter<DreamItemViewHolder> {
 
     override fun onBindViewHolder(holder: DreamItemViewHolder, position: Int) {
         holder?.bind(items[position], context)
+        holder?.civ.tag = object : Target {
+            override fun onPrepareLoad(placeHolderDrawable: Drawable?) {
+                Log.d("image", "Prepare Load")
+                holder.civ.setImageResource(R.mipmap.ic_launcher)
+            }
+
+            override fun onBitmapFailed(e: Exception?, errorDrawable: Drawable?) {
+                Log.d("image", "Failed")
+    //                civ?.setImageResource(R.mipmap.ic_launcher)
+                holder.civ.setImageResource(R.mipmap.ic_launcher)
+            }
+
+            override fun onBitmapLoaded(bitmap: Bitmap?, from: Picasso.LoadedFrom?) {
+                Log.d("image", "Imaged Loaded")
+                holder.civ.setImageDrawable(BitmapDrawable(context.resources, bitmap))
+            }
+        }
+
+        Picasso.get().load(items[position].button_img).resize(92, 92).into(holder.civ.tag as Target)
+
         holder?.itemView.setOnClickListener {
             detailIntent.putExtra("date", items[position].date)
             detailIntent.putExtra("during", items[position].during)
@@ -50,5 +85,17 @@ class DreamItemAdapter : RecyclerView.Adapter<DreamItemViewHolder> {
 
     }
 
-
+//    override fun onBindViewHolder(holder: DreamItemViewHolder, position: Int, payloads: MutableList<Any>) {
+//        if(payloads.isEmpty()) {
+//            super.onBindViewHolder(holder, position, payloads)
+//        }else{
+//            for (payload in payloads) {
+//                if (payload is String) {
+//                    if (TextUtils.equals(payload, "update") && holder is DreamItemViewHolder) {
+//                        notifyItemChanged(position)
+//                    }
+//                }
+//            }
+//        }
+//    }
 }
