+++ app/src/main/res/drawable/bg_delete_dream.xml
@@ -3,9 +3,9 @@
     <item>
         <shape>
             <gradient
-                android:angle="90"
-                android:startColor="#6dffffff"
-                android:endColor="#6dffffff"/>
+                    android:angle="90"
+                    android:startColor="#6dffffff"
+                    android:endColor="#6dffffff" />
             <corners android:radius="21dp"/>
         </shape>
     </item>
